+++ component_registry_bindings/constants.py
@@ -1,7 +1,14 @@
+"""
+component-registry-bindings constants
+"""
+
 from typing import Dict, List
 
 COMPONENT_REGISTRY_API_VERSION: str = "v1"
-COMPONENT_REGISTRY_BINDINGS_USERAGENT: str = "component-registry-bindings-1.3.4"
+COMPONENT_REGISTRY_BINDINGS_VERSION: str = "1.3.4"
+COMPONENT_REGISTRY_BINDINGS_USERAGENT: str = (
+    f"component-registry-bindings-{COMPONENT_REGISTRY_BINDINGS_VERSION}"
+)
 COMPONENT_REGISTRY_BINDINGS_API_PATH: str = (
     f".bindings.python_client.api.{COMPONENT_REGISTRY_API_VERSION}"
 )
