+++ component_registry_bindings/exceptions.py
@@ -1,0 +1,11 @@
+"""
+component-registry-bindings exceptions
+"""
+
+
+class ComponentRegistryBindingsException(Exception):
+    """Base component-registry-bindings exception"""
+
+
+class OperationUnsupported(ComponentRegistryBindingsException):
+    """Session operation is unsupported exception"""
